rename GeoSeries to GeometrySeries
